# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "AMAZING SOULS API",
        "description": "API documentation for AMAZING SOULS API",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/auth/verify-login-otp": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "Verify login OTP and complete authentication",
                "operationId": "verifyLoginOtp",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "otp_challenge_token",
                                    "otp"
                                ],
                                "properties": {
                                    "otp_challenge_token": {
                                        "type": "string",
                                        "example": "challenge-token-from-login-response"
                                    },
                                    "otp": {
                                        "type": "string",
                                        "example": "123456"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "OTP verified and login successful"
                    },
                    "422": {
                        "description": "Invalid OTP code"
                    },
                    "400": {
                        "description": "Invalid or expired challenge"
                    },
                    "429": {
                        "description": "Too many OTP attempts"
                    }
                }
            }
        },
...
        "/api/auth/forgot-password": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "Request password reset link",
                "operationId": "forgotPassword",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "email"
                                ],
                                "properties": {
                                    "email": {
                                        "type": "string",
                                        "format": "email",
                                        "example": "john@example.com"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Reset link requested"
                    }
                }
            }
        },
        "/api/auth/reset-password": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "Reset password with token",
                "operationId": "resetPassword",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "email",
                                    "token",
                                    "password",
                                    "password_confirmation"
                                ],
                                "properties": {
                                    "email": {
                                        "type": "string",
                                        "format": "email",
                                        "example": "john@example.com"
                                    },
                                    "token": {
                                        "type": "string",
                                        "example": "reset-token-here"
                                    },
                                    "password": {
                                        "type": "string",
                                        "format": "password",
                                        "example": "newsecret1234"
                                    },
                                    "password_confirmation": {
                                        "type": "string",
                                        "format": "password",
                                        "example": "newsecret1234"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Password reset successful"
                    },
                    "400": {
                        "description": "Invalid or expired token"
                    }
                }
            }
        },
...
        "/api/auth/verify-token": {
            "get": {
                "tags": [
                    "Auth"
                ],
                "summary": "Verify current auth token",
                "operationId": "verifyToken",
                "responses": {
                    "200": {
                        "description": "Token is valid"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/auth/two-factor": {
            "put": {
                "tags": [
                    "Auth"
                ],
                "summary": "Enable or update login OTP for two-factor authentication",
                "operationId": "enableOrUpdateTwoFactor",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "otp",
                                    "otp_confirmation"
                                ],
                                "properties": {
                                    "otp": {
                                        "type": "string",
                                        "example": "123456"
                                    },
                                    "otp_confirmation": {
                                        "type": "string",
                                        "example": "123456"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Two-factor updated successfully"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "422": {
                        "description": "Validation error"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Auth"
                ],
                "summary": "Disable two-factor authentication",
                "operationId": "disableTwoFactor",
                "responses": {
                    "200": {
                        "description": "Two-factor disabled successfully"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "sanctum": {
                "type": "http",
                "description": "Use a Sanctum personal access token. Swagger will send it as Authorization: Bearer {token}.",
                "bearerFormat": "Sanctum",
                "scheme": "bearer"
            }
        }
    },
...
}